Split web workspace session package
