+++ apps/web/src/appData/session/activation/workspaceActivationHelpers.ts
@@ -1,0 +1,1 @@
+export const defaultWorkspaceName: string = "Personal";
