+++ apps/web/src/appData/session/cloud/workspaceSessionCloud.ts
@@ -1,6 +1,6 @@
-import { getStableInstallationId } from "../../clientIdentity";
-import type { LocalBrowserDataCleanupReason } from "../../accountDeletion";
-import type { CloudSettings, SessionInfo } from "../../types";
+import { getStableInstallationId } from "../../../clientIdentity";
+import type { LocalBrowserDataCleanupReason } from "../../../accountDeletion";
+import type { CloudSettings, SessionInfo } from "../../../types";
 
 export function buildLinkingReadyCloudSettings(session: SessionInfo): CloudSettings {
   return {
