+++ apps/web/src/appData/session/lifecycle/workspaceLifecycleHelpers.ts
@@ -1,7 +1,3 @@
-import type { TranslationKey } from "../../i18n";
-import type { WorkspaceSummary } from "../../types";
-
-export const defaultWorkspaceName: string = "Personal";
 export const resumeRetryDelayMs: number = 750;
 export const resumeRetryCount: number = 2;
 
@@ -10,23 +6,6 @@ const sessionAccountSwitchErrorName = "SessionAccountSwitchError";
 export type SessionAccountSwitchError = Error & Readonly<{
   name: typeof sessionAccountSwitchErrorName;
 }>;
-
-export function replaceWorkspaceSummary(
-  workspaces: ReadonlyArray<WorkspaceSummary>,
-  workspace: WorkspaceSummary,
-): ReadonlyArray<WorkspaceSummary> {
-  let didReplace = false;
-  const nextWorkspaces = workspaces.map((currentWorkspace) => {
-    if (currentWorkspace.workspaceId !== workspace.workspaceId) {
-      return currentWorkspace;
-    }
-
-    didReplace = true;
-    return workspace;
-  });
-
-  return didReplace ? nextWorkspaces : [...workspaces, workspace];
-}
 
 export function consumeLoggedOutMarker(): boolean {
   const url = new URL(window.location.href);
@@ -38,10 +17,6 @@ export function consumeLoggedOutMarker(): boolean {
   const nextUrl = `${url.pathname}${url.search}${url.hash}`;
   window.history.replaceState({}, document.title, nextUrl);
   return true;
-}
-
-export function createRemoteActionLockedError(t: (key: TranslationKey) => string): Error {
-  return new Error(t("app.sessionRestoringActionLocked"));
 }
 
 export function createSessionAccountSwitchError(errorMessage: string): SessionAccountSwitchError {
