+++ apps/web/src/appData/session/useWorkspaceSession.test.tsx
@@ -10,14 +10,14 @@ import {
 import { setNavigationHandlerForTests, resetApiClientStateForTests } from "../../api";
 import { INSTALLATION_ID_STORAGE_KEY } from "../../clientIdentity";
 import { LOCALE_PREFERENCE_STORAGE_KEY } from "../../i18n/runtime";
-import { loadWarmStartSnapshot, WARM_START_SNAPSHOT_STORAGE_KEY } from "./warmStart";
+import { loadWarmStartSnapshot, WARM_START_SNAPSHOT_STORAGE_KEY } from "./activation/warmStart";
 import { useWorkspaceSession } from "./useWorkspaceSession";
-import { captureWorkspaceTransitionError } from "./workspaceSessionObservation";
+import { captureWorkspaceTransitionError } from "./observation/workspaceSessionObservation";
 import { putCloudSettings, loadCloudSettings } from "../../localDb/sync/cloudSettings";
 import type { CloudSettings, SessionInfo, WorkspaceSummary } from "../../types";
 import type { TranslationKey } from "../../i18n";
 import type { SessionLoadState } from "../context/types";
-import type { SessionVerificationState } from "./warmStart";
+import type { SessionVerificationState } from "./workspaceSessionTypes";
 import { clearWebSyncCache } from "../../localDb/cache";
 
 const observabilityMocks = vi.hoisted(() => ({
